moved tui widgets related code to `pkg` folder
