+++ pkg/tui/widgets/focusable.go
