+++ pkg/tui/widgets/list/styles.go
